+++ packages/apptive_grid_form/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.10.0-alpha.5
+* Add `description` to forms
+
 ## 0.10.0-alpha.4
 * Update Core Package to 0.10.0-alpha.4 to include breaking changes to GridField